+++ src/main/java/model/coordinatecalculation/strategy/TriangleAreaCalculationStrategy.java
@@ -1,4 +1,8 @@
-package model;
+package model.coordinatecalculation.strategy;
+
+import model.CoordinateCalculationResult;
+import model.Coordinates;
+import model.Point;
 
 import java.util.List;
